+++ src/test/inspection/common.ts
@@ -2,9 +2,20 @@
 // Licensed under the MIT license.
 
 import "mocha";
+import { Inspection } from "../..";
 import { Option, ResultKind } from "../../common";
 import { Lexer, LexerSnapshot, TriedLexerSnapshot } from "../../lexer";
-import { IParserState, IParserStateUtils, ParseOk, Parser, ParseError, TriedParse } from "../../parser";
+import { IParserState, IParserStateUtils, ParseError, ParseOk, Parser, TriedParse } from "../../parser";
+
+export function expectParseOkInspection(text: string, position: Inspection.Position): Inspection.TriedInspection {
+    const parseOk: ParseOk = expectParseOk(text);
+    return Inspection.tryFrom(position, parseOk.nodeIdMapCollection, parseOk.leafNodeIds);
+}
+
+export function expectParseErrInspection(text: string, position: Inspection.Position): Inspection.TriedInspection {
+    const parserError: ParseError.ParseError = expectParseErr(text);
+    return Inspection.tryFrom(position, parserError.context.nodeIdMapCollection, parserError.context.leafNodeIds);
+}
 
 export function expectParseErr(text: string): ParseError.ParseError {
     const triedParse: TriedParse = expectTriedParse(text);
